+++ packages/automation-contracts/scheduler/hardhat.config.js
@@ -25,13 +25,6 @@ module.exports = {
             url: "http://127.0.0.1:8545/",
             chainId: 31337,
         },
-        mumbai: {
-            url: process.env.MUMBAI_URL || "",
-            accounts:
-                process.env.PRIVATE_KEY !== undefined
-                    ? [process.env.PRIVATE_KEY]
-                    : [],
-        },
         polygon: {
             url: process.env.POLYGON_URL || "",
             accounts: